git commit

+++ fe.cpp
@@ -2,10 +2,10 @@
 #include <cmath>
 
 
-int f(x);
+double f(x);
 int main()
 {
     int x=2;
-    int res = f(int x);
+    double res = f(int x);
     return 0;
 }
